+++ lib/utils/consts.dart
@@ -3,3 +3,4 @@ const enableNotificationsFlag = 'enable_notifications';
 const showBrightSchemeFlag = 'show_bright_scheme';
 const allowInvalidCertFlag = 'allow_invalid_cert';
 const enableSyncFlag = 'enable_sync';
+const recordLocationFlag = 'record_location';