+++ lib/widgets/habits/habit_category.dart
@@ -71,7 +71,7 @@ class SelectCategoryWidget extends StatelessWidget {
                                   Navigator.pop(context);
                                 },
                                 title: category.name,
-                                leading: CategoryColorIcon(
+                                leading: ColorIcon(
                                   colorFromCssHex(category.color),
                                 ),
                               ),
@@ -97,8 +97,9 @@ class SelectCategoryWidget extends StatelessWidget {
               decoration: inputDecoration(
                 labelText:
                     categoryUndefined ? '' : localizations.habitCategoryLabel,
+                semanticsLabel: 'Select category',
               ).copyWith(
-                icon: CategoryColorIcon(
+                icon: ColorIcon(
                   category != null
                       ? colorFromCssHex(category.color)
                       : styleConfig().secondaryTextColor.withOpacity(0.2),